brand tracker/requirements.txt
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Cython==0.29.23
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dropbox==11.23.0
Flask==2.0.2
Flask-Compress==1.10.1
gensim==4.1.2
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
heroku3==5.1.4
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
kaggle==1.5.12
MarkupSafe==2.0.1
more-itertools==8.12.0
mysql-connector-python==8.0.26
nltk==3.6.5
numpy==1.21.3
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.3.4
plotly==5.3.1
ply==3.11
praw==7.5.0
prawcore==2.3.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pyparsing==3.0.5
python-dateutil==2.8.2
python-dotenv==0.19.1
python-slugify==5.0.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.1
stone==3.2.1
tenacity==8.0.1
text-unidecode==1.3
textblob==0.17.1
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==3.10.0.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
xgboost==1.5.1
XlsxWriter==3.0.1
zipp==3.6.0
